+++ lib/constant/record_date.dart
@@ -1,0 +1,3 @@
+class RecordDate {
+  static final DateTime termsLastUpdateDay = DateTime(2022, 4, 1);  // 利用規約の最終更新日
+}
